Undo split code
Why did we have sim2.c and sim.c if sim.c is just old code? This is what
github's for!
